ps4-rag-qa-starter/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.43.3
accelerate==0.33.0
torch>=2.2.0
numpy==1.26.4
pandas==2.2.2
pypdf==4.3.1
scikit-learn==1.5.1
pytest==8.2.0
